ref(crons): Move volume recording logic to system_incidents module
