+++ src/sentry/monitors/consumers/monitor_consumer.py
@@ -28,7 +28,7 @@ from sentry.constants import DataCategory, ObjectStatus
 from sentry.db.postgres.transactions import in_test_hide_transaction_boundary
 from sentry.killswitches import killswitch_matches_context
 from sentry.models.project import Project
-from sentry.monitors.clock_dispatch import try_monitor_clock_tick, update_check_in_volume
+from sentry.monitors.clock_dispatch import try_monitor_clock_tick
 from sentry.monitors.constants import PermitCheckInStatus
 from sentry.monitors.logic.mark_failed import mark_failed
 from sentry.monitors.logic.mark_ok import mark_ok
@@ -64,6 +64,7 @@ from sentry.monitors.processing_errors.errors import (
     ProcessingErrorType,
 )
 from sentry.monitors.processing_errors.manager import handle_processing_errors
+from sentry.monitors.system_incidents import update_check_in_volume
 from sentry.monitors.types import CheckinItem
 from sentry.monitors.utils import (
     get_new_timeout_at,
